refactor: update to latest Vitess version and clean code
